refactoring component out
